Update atividade 11

+++ src/app/user/user.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { NgForm } from '@angular/forms';
+import { BehaviorSubject, ConnectableObservable } from 'rxjs';
 import { User } from '../model/User';
+import { UserService } from '../service/user.service';
 
 @Component({
   selector: 'app-user',
@@ -8,41 +10,56 @@ import { User } from '../model/User';
   styleUrls: ['./user.component.css']
 })
 export class UserComponent implements OnInit {
+  @ViewChild('form') form!: NgForm;
   user!: User;
   users: User[];
+  usersModify?: User[];
+  userRepassword: string = '';
+  valorArray!: BehaviorSubject<number>;
 
-  formulario: any;
-  constructor() {
+  constructor( private userService : UserService) {
     this.users = []
+    this.getAll();
   }
 
   ngOnInit(): void {
-     this.getAll();
-     this.formulario = new FormGroup({
-      id: new FormControl(),
-      username: new FormControl(),
-      email: new FormControl(),
-      password: new FormControl(),
-      permission: new FormControl()
-     })
-
-
+    this.user = new User('','','','')
   }
 
   onSubmit() : void {
-    this.formulario.value.id = String(Math.round(Math.random() * 100));
-    const usuario : User = this.formulario.value;
-    this.users.push(usuario);
-    localStorage.setItem('user', JSON.stringify(this.users))
-    this.formulario.reset();
+    if(this.isExist(this.user.username)){
+        this.users =  this.users.filter((u) => {
+          return u.username?.valueOf() != this.user.username?.valueOf();
+        });
+        this.users.push(this.user);
+        localStorage.setItem('user', JSON.stringify(this.users))
+        this.form.reset();
+        this.getAll();
+    }else{
+      this.users.push(this.user);
+      localStorage.setItem('user', JSON.stringify(this.users))
+      this.form.reset();
+      this.getAll();
+    }
+
   }
 
   getAll() : void{
-    if(localStorage.getItem('user')){
-      this.users = JSON.parse(localStorage.getItem('user')!);
-    }else{
-      this.users = [];
-    }
+     this.users = JSON.parse(localStorage.getItem('user')!);
+  }
 
+  onUpdate(user: User) : void {
+    const usuario : User = user;
+    this.user = usuario;
+  }
+
+  isExist(value: string): boolean {
+
+    for (let u of this.users) {
+      if (u.username?.valueOf() == value?.valueOf()) {
+          return true;
+      }
+    }
+    return false;
   }
 }
